+++ src/server/scripts/WrathOfTheLichKing/Raids/IcecrownCitadel/boss_lady_deathwhisper.cpp
@@ -296,7 +296,7 @@ class boss_lady_deathwhisper : public CreatureScript
             void DamageTaken(Unit* /*damageDealer*/, uint32& uiDamage)
             {
                 // phase transition
-                if (events.GetPhaseMask() & PHASE_ONE_MASK && uiDamage > me->GetPower(POWER_MANA))
+                if (events.GetPhaseMask() & PHASE_ONE_MASK && int32(uiDamage) > me->GetPower(POWER_MANA))
                 {
                     //Talk(SAY_PHASE_2);
                     //Talk(EMOTE_PHASE_2);
